requirements_streamlit.txt
streamlit==1.26.0
tensorflow-cpu==2.15.0
Pillow==10.0.0
numpy==1.24.3
opencv-python-headless==4.8.1.78
scikit-learn==1.3.0
protobuf==4.25.8
typing_extensions==4.14.1
python-dateutil==2.9.0.post0
